feat(states): add new state types
Added `ButtonState` and `SwitchState` to the list of supported state types in `esphome_controller.dart`. Also created a new file `state_button.dart` and `state_switch.dart` to define the new state types. Additionally, created a new abstract class `EspHomeElement` in `esphome_element.dart` to be used as a base class for all state types. This will make it easier to add new state types in the future.